CleanHistograms

Documentation for CleanHistograms.

CleanHistograms.binweave

CleanHistograms.binweave!

CleanHistograms.cleanhist

CleanHistograms.interleave

CleanHistograms.interleave!

CleanHistograms.quickhist

CleanHistograms.quickhist!

CleanHistograms.binweave! — Method

binweave!(c,a)

In-place version of binweave, that overwrites c with woven bin edges.

CleanHistograms.binweave — Method

binweave(a)

Interleave the sequential values of a (<:AbstractVector) using the first and last values only once. To be used with interleave to make plotting-ready histograms.

Example

julia> binweave([1,2,3])
4-element Vector{Int64}:
1
2
2
3

CleanHistograms.cleanhist — Method

cleanhist(x; nbins=32, scooch_nbins=2)

Calculates a histogram with extra (0 count) bins to buffer the edges and make it look nice and clean. 🧼

Optionally specify the number of histogram bins (default: 2⁵ bins) and the number of buffering bins scooch. (Total bins = nbins + scoochbins)

Returns a NamedTuple with x and y values of histogram.

CleanHistograms.interleave! — Method

interleave!(c,a)

In-place version of interleave, that fills c with pairs of each index in a.

CleanHistograms.interleave — Method

interleave(a)

Interleave the sequential values of a vector a so that each value occurs twice in a vector of length(a)*2

Example

julia> interleave([1,2,3])
6-element Vector{Int64}:
1
1
2
2
3
3

CleanHistograms.quickhist! — Method

quickhist!(A, x, binedges)

In-place form of quickhist that overwrites A, which must have a length equal binedges-1.

CleanHistograms.quickhist — Method

quickhist(x, binedges)

Quickly calculate histogram counts from collection x from an Abstract Range of binedges. Note that any values in x that fall outside of binedges will not be counted. Use at your own risk!